Split type-name resolution out of PInvokeGenerator.cs
Move the contiguous type-name and type-size resolution helpers
(GetTargetTypeName, GetTypeName, GetTypeNameForPointeeType, and
GetTypeSize) into a dedicated PInvokeGenerator.TypeResolution.cs partial.
This trims the main partial by ~1150 lines and groups the related type
handling together. Pure code move; no behavior change.
